first/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
djangorestframework==3.15.1
docutils==0.21.1
drf-spectacular==0.27.2
flake8==7.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.2.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
